feat(codegen)：支持自定义父目录 (#33)
* feat(doc): readme 修改

* feat(codegen): 路径拼接方式改为使用 path.Join

* feat(codegen): path.Join改为 filepath.Join

* feat(gast): 向 map 中追加 kv

+++ codegen/mysql.go
@@ -84,8 +84,8 @@ var mysqlDefaultColumnTypeMap = map[string]string{
 	"mediumint":  "int32",
 	"int":        "int32",
 	"integer":    "int32",
-	"bigint":     "uint64",
-	"float":      "float32",
+	"bigint":     "uint",
+	"float":      "float64",
 	"double":     "float64",
 	"decimal":    "string", // 或者使用 "big.Rat" 或 "float64"，取决于精度需求
 	"date":       "time.Time",
